Create Password-DB-Schema
This schema creates a table named USERS with columns for id, username, and password. The id column is an auto-incrementing primary key. The username column stores the username of the user, and the password column stores the hashed password. The username column is set to be unique to ensure each username is unique in the database. Finally, the character set and collation for the table are set to utf8mb4 to support storing a wide range of characters.